# executor: tiny optimize slow_query performance and add related runtime stats (#20200)