cherry-pick #20200 to release-4.0

---

<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Tiny optimization for query `SLOW_QUERY` when there are too many slow log file (more than 50G).

1. Port optimization from, https://github.com/pingcap/sysutil/pull/19/files#r451300500 
2. Avoid `filepath.Walk` since it will deep traverse every file.

Bbenchmark for `point-get` `SLOW_QUERY` when there are 50G slow-log (slow log file store in mechanical hard disk, not in nvme).
```sql
select time from information_schema.SLOW_QUERY where time = "2020-09-24 17:33:00.909273";
```

**before**

spend `490ms`

**this PR**

spend `130ms`

3. add runtime stats for slow_query:

```sql
> explain analyze select count(*) from `SLOW_QUERY` where time >= "2020-09-24 17:33:00.909273" and time < now();
+-------------------+----------+---------+------+------------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------+---------------+------+
| id                | estRows  | actRows | task | access object    | execution info                                                                                                                                                                             | operator info                                                              | memory        | disk |
+-------------------+----------+---------+------+------------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------+---------------+------+
| HashAgg_6         | 1.00     | 1       | root |                  | time:115.215614ms, loops:2, PartialConcurrency:5, FinalConcurrency:5                                                                                                                       | funcs:count(1)->Column#57                                                  | 417.078125 KB | N/A  |
| └─MemTableScan_11 | 10000.00 | 114     | root | table:SLOW_QUERY | time:115.135014ms, loops:4, initialize: 4.509084ms, read_file: 107.308575ms, parse_log: {time:48.243624ms, concurrency:15}, total_file: 139, read_file: 1, read_size: 41.76433753967285 MB | start_time:2020-09-24 17:33:00.909273, end_time:2020-09-24 20:07:02.999000 | N/A           | N/A  |
+-------------------+----------+---------+------+------------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------+---------------+------+
```

### What is changed and how it works?


### Related changes

- Need to cherry-pick to the release branch

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Manual test

Side effects

- No

### Release note <!-- bugfixes or new feature need a release note -->

- No.